requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.1.0
PyPDF2==3.0.1
pdf2docx==0.5.6
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
img2pdf==0.5.1
pypandoc==1.12
rembg==2.0.56
opencv-python==4.8.1.78
numpy==1.24.3
moviepy==1.0.3
pydub==0.25.1
PyMuPDF==1.23.8
reportlab==4.0.7
Werkzeug==3.0.1
python-magic==0.4.27
vtracer==0.6.11
imagetracerpy==1.0.2